% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve dataset from statnipokladna}
\usage{
get_dataset(dataset_id, year = 2018, month = 12, force_redownload = F)
}
\arguments{
\item{dataset_id}{A dataset ID. See \code{id} column in \code{sp_datasets} for a list of available codelists.}

\item{year}{year, numeric, 2015-2018 for some datasets, 2010-2018 for others.}

\item{month}{month, numeric. Must be 3, 6, 9 or 12.}

\item{force_redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}
}
\value{
character string with complete paths to downloaded files.
}
\description{
Downloads and unzips files for a given dataset.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Core workflow: 
\code{\link{get_codelist}()},
\code{\link{get_table}()}
}
\concept{Core workflow}
