% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_doc}
\alias{get_dataset_doc}
\title{Get dataset documentation}
\usage{
get_dataset_doc(dataset_id, destdir = ".", download = T)
}
\arguments{
\item{dataset_id}{dataset ID. See \code{sp_datasets}.}

\item{destdir}{character. Where the file should be written. Defaults to working directory.}

\item{download}{Whether to download (the default) or open link in browser.}
}
\value{
character (link) if download = TRUE, nothing otherwise.
}
\description{
Downloads XLS file with dataset documentation, or opens link to this file in browser.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Utilities: 
\code{\link{get_codelist_viewer}()}
}
\concept{Utilities}
