% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp_datasets}
\alias{sp_datasets}
\title{List of available datasets}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{\code{id}}{character. Dataset ID, used as \code{dataset_id} argument to \code{sp_get_dataset}.}
\item{\code{name}}{character. Dataset name, mostly corresponds to title on the statnipokladna GUI.}
}
}
\usage{
sp_datasets
}
\description{
Contains IDs and names of all available datasets that can be retrieved by get_dataset.
}
\details{
See \url{https://monitor.statnipokladna.cz/datovy-katalog/transakcni-data} for a more detailed descriptions
of the datasets.
}
\seealso{
Other Lists of available entities: 
\code{\link{sp_codelists}},
\code{\link{sp_tables}}
}
\concept{Lists of available entities}
\keyword{datasets}
