% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{sp_get_codelist_url}
\alias{sp_get_codelist_url}
\title{Get URL of a given codelist}
\usage{
sp_get_codelist_url(codelist_id, check_if_exists = TRUE)
}
\arguments{
\item{codelist_id}{DESCRIPTION.}

\item{check_if_exists}{Whether to check that the URL works (HTTP 200).}
}
\value{
character vector of length one containing URL
}
\description{
This is normally called inside \code{sp_get_codelist()} but can be used separately if
finer-grained control of intermediate outputs is needed, e.g. in a {targets} workflow.
}
\examples{
\dontrun{
sp_get_codelist_url("ucjed", FALSE)
if(FALSE) sp_get_codelist_url("ucjed_wrong", TRUE) # fails, invalid codelist
}
}
