% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{sp_get_codelist_viewer}
\alias{sp_get_codelist_viewer}
\title{Get/open URL of codelist viewer}
\usage{
sp_get_codelist_viewer(codelist_id, open = TRUE)
}
\arguments{
\item{codelist_id}{A codelist ID. See \code{id} column in \code{sp_codelists} for a list of available codelists.}

\item{open}{Whether to open URL in browser. Defaults to TRUE.}
}
\value{
a URL, character vector of length one.
}
\description{
Returns a URL for the online codelist browser in monitor.statnipokladna.cz and opens it in browser if open = TRUE.
}
\seealso{
Other Utilities: 
\code{\link{sp_get_dataset_doc}()}
}
\concept{Utilities}
\keyword{internal}
