% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{sp_list_datasets}
\alias{sp_list_datasets}
\title{List all files currently available from the data provider}
\format{
A data frame with these variables:
\describe{
\item{\code{table}}{character. Table name incl. period (long name, does not correspond to dataset label in \code{sp_tables}).}
\item{\code{dataset}}{character. Dataset (long name, does not correspond to dataset label in \code{sp_datasets}).}
\item{\code{start}}{date. Start date of temporal coverage for this file.}
\item{\code{end}}{date End date of temporal coverage for this file.}
\item{\code{filetype}}{character. Filetyp. Currently 'csv' for all files.}
\item{\code{compression}}{character. Type of compression. Currently 'zip' for all files.}
\item{\code{url}}{character. Link to downloadable file/archive.}
\item{\code{doc}}{character. Link to documantation Currently empty as not provided by endpoint.}
\item{\code{schema}}{character. Link to schema. Currently empty as not provided by endpoint.}
}
}
\usage{
sp_list_datasets()
}
\value{
A \link[tibble:tibble-package]{tibble} with one row per downloadable file
}
\description{
Queries the SPARQL endpoint at \url{https://opendata.mfcr.cz/lod/monitor}
and
}
\examples{
\donttest{
sp_get_catalogue()
}
}
\seealso{
Other Lists of available entities: 
\code{\link{sp_codelists}},
\code{\link{sp_datasets}},
\code{\link{sp_tables}}
}
\concept{Lists of available entities}
