% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{sp_tables}
\alias{sp_tables}
\title{List of available tables}
\format{A data frame with 2 rows and 4 variables:
\describe{
\item{\code{id}}{integer Table id, used as \code{id} argument to \code{get_table}. Comprises report number and table number.}
\item{\code{table_code}}{character. Table code, should be human readable.}
\item{\code{table_num}}{integer Table number.}
\item{\code{report_num}}{integer Number of report (vykaz) containing the table.}
}}
\usage{
sp_tables
}
\description{
Contains IDs and names of all available tables that can be
retrieved by get_table. Look inside the XLS documentation for each dataset at \url{https://monitor.statnipokladna.cz/2019/zdrojova-data/transakcni-data}
to see more detailed descriptions. Note that tables do not correspond to the tabulka/\code{vtab} attribute of the tables, they represent files inside datasets.
}
\seealso{
Other Lists of available entities: 
\code{\link{sp_codelists}},
\code{\link{sp_datasets}}
}
\concept{Lists of available entities}
\keyword{datasets}
