% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnipokladna-package.R
\docType{package}
\name{statnipokladna-package}
\alias{statnipokladna}
\alias{statnipokladna-package}
\title{statnipokladna: Use Data from the Czech Public Finance Database}
\description{
This package provides programmatic access to
    data from the Czech public budgeting and accounting database, Státní
    pokladna.
}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com}

}
\keyword{internal}
