% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{add_codelist}
\alias{add_codelist}
\title{Deprecated: Add codelist data to downloaded data}
\usage{
add_codelist(
  data,
  codelist = NULL,
  period_column = .data$vykaz_date,
  redownload = FALSE,
  dest_dir = NULL
)
}
\arguments{
\item{data}{a data frame returned by \code{sp_get_table()}.}

\item{codelist}{The codelist to add. Either a character vector of length one (see \code{sp_tables} for possible values), or a data frame returned by \code{sp_get_codelist()}.}

\item{period_column}{Unquoted column name of column identifying the data period in \code{data}. Leave to default if you have not changed the \code{data} object returned by \code{sp_get_table()}.}

\item{redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}
}
\value{
A \link[tibble:tibble-package]{tibble} of same length as \code{data}, with added columns from \code{codelist}. See Details.
}
\description{
Deprecated, use \code{sp_add_codelist()} instead.\cr\cr
\lifecycle{deprecated}
}
\seealso{
Other Core workflow: 
\code{\link{get_codelist}()},
\code{\link{sp_add_codelist}()},
\code{\link{sp_get_codelist}()},
\code{\link{sp_get_dataset}()},
\code{\link{sp_get_table}()}
}
\concept{Core workflow}
