% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{get_codelist}
\alias{get_codelist}
\title{Get codelist}
\usage{
get_codelist(codelist_id, n = NULL, force_redownload = F)
}
\arguments{
\item{codelist_id}{A codelist ID. See \code{id} column in \code{sp_codelists} for a list of available codelists.}

\item{n}{Number of rows to return. Default (NULL) means all. Useful for quickly inspecting a codelist.}

\item{force_redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}
}
\value{
A tibble
}
\description{
Downloads and processes codelist identified by \code{codelist_id}. See \code{sp_codelists} for a list of
of available codelists with their IDs and names.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Core workflow: 
\code{\link{get_dataset}()},
\code{\link{get_table}()}
}
\concept{Core workflow}
