% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{get_codelist}
\alias{get_codelist}
\title{Get codelist}
\usage{
get_codelist(codelist)
}
\arguments{
\item{codelist}{A codelist ID. See \code{id} column in \code{sp_codelists} for a list of available codelists.}
}
\value{
A tibble
}
\description{
Downloads and processes codelist identified by \code{codelist}. See \code{sp_codelists} for a list of
of available codelists with their IDs and names.
}
\examples{
# ADD_EXAMPLES_HERE
}
