% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{get_codelist}
\alias{get_codelist}
\title{Deprecated: Get codelist}
\usage{
get_codelist(codelist_id, n = NULL, dest_dir = NULL, redownload = FALSE)
}
\arguments{
\item{codelist_id}{A codelist ID. See \code{id} column in \code{sp_codelists} for a list of available codelists.}

\item{n}{Number of rows to return. Default (NULL) means all. Useful for quickly inspecting a codelist.}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}

\item{redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Deprecated: use \code{sp_get_codelist()}\cr\cr
\lifecycle{deprecated}
}
\seealso{
Other Core workflow: 
\code{\link{add_codelist}()},
\code{\link{sp_add_codelist}()},
\code{\link{sp_get_codelist}()},
\code{\link{sp_get_dataset}()},
\code{\link{sp_get_table}()}
}
\concept{Core workflow}
