% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve dataset from statnipokladna}
\usage{
get_dataset(
  dataset_id,
  year = 2019,
  month = 12,
  dest_dir = tempdir(),
  redownload = F
)
}
\arguments{
\item{dataset_id}{A dataset ID. See \code{id} column in \code{sp_datasets} for a list of available codelists.}

\item{year}{year, numeric, 2015-2018 for some datasets, 2010-2018 for others. Defaults to 2018.}

\item{month}{month, numeric. Must be 3, 6, 9 or 12. Defaults to 12.}

\item{force_redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}
\item{dest_dir}{character. Directory in which downloaded files will be stored. Defaults to \code{tempdir()}.}
}
\value{
character string with complete paths to downloaded files.
}
\description{
Downloads and unzips files for a given dataset.
}
\details{
Files are stored in a temp folder as determined by \code{tempdir()} and further sorted into
subdirectories by dataset, year and month. They persist per session to avoid redownloads.
}
\examples{
\dontrun{
budget_latest <- get_dataset("finu")
budget_2018 <- get_dataset("finu", 2018)
budget_mid2018 <- get_dataset("finu", 2018, 6)
budget_multiyear <- get_dataset("finu", 2012:2018, 9)
budget_multihalfyears <- get_dataset("finu", 2012:2018, c(6, 12))
}
}
\seealso{
Other Core workflow: 
\code{\link{add_codelist}()},
\code{\link{get_codelist}()},
\code{\link{get_table}()}
}
\concept{Core workflow}
