% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Deprecated: Retrieve and read dataset from statnipokladna}
\usage{
get_dataset(dataset_id, year, month = 12, dest_dir = NULL, redownload = FALSE)
}
\arguments{
\item{dataset_id}{A dataset ID. See \code{id} column in \code{sp_datasets} for a list of available datasets}

\item{year}{year, numeric vector of length <= 1 (can take multiple values), 2015-2019 for some datasets, 2010-2020 for others. Defaults to 2018.
(see Details for how to work with data across time periods.)}

\item{month}{month, numeric vector of length <= 1 (can take multiple values). Must be between 1 and 12. Defaults to 12.
(see Details for how to work with data across time periods.)}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}

\item{redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}
}
\value{
character (link) if download = TRUE, nothing otherwise.
}
\description{
Deprecated, use \code{sp_get_dataset()} instead.\cr\cr
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
