% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Deprecated: Retrieve and read dataset from statnipokladna}
\usage{
get_dataset(
  dataset_id,
  year = 2019,
  month = 12,
  dest_dir = NULL,
  redownload = FALSE
)
}
\arguments{
\item{dataset_id}{A dataset ID. See \code{id} column in \code{sp_datasets} for a list of available codelists.}

\item{year}{year, numeric, 2015-2018 for some datasets, 2010-2019 for others. Defaults to 2018.
(see Details for how to work with data across time periods.)}

\item{month}{month, numeric. Must be between 1 and 12. Defaults to 12.
(see Details for how to work with data across time periods.)}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}

\item{redownload}{Redownload even if file has already been downloaded? Defaults to FALSE.}
}
\value{
character (link) if download = TRUE, nothing otherwise.
}
\description{
Deprecated, use \code{sp_get_dataset()} instead.\cr\cr
\lifecycle{deprecated}
}
\seealso{
Other Utilities: 
\code{\link{sp_get_codelist_viewer}()},
\code{\link{sp_get_dataset_doc}()}
}
\concept{Utilities}
