% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_doc}
\alias{get_dataset_doc}
\title{Deprecated: Get dataset documentation}
\usage{
get_dataset_doc(dataset_id, dest_dir = ".", download = TRUE)
}
\arguments{
\item{dataset_id}{dataset ID. See \code{sp_datasets}.}

\item{dest_dir}{character. Where the file should be written. Defaults to working directory.}

\item{download}{Whether to download (the default) or open link in browser.}
}
\value{
a tibble
}
\description{
Deprecated, use \code{sp_get_dataset_doc()} instead.
}
\details{
\lifecycle{soft-deprecated}
}
