% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_doc}
\alias{get_dataset_doc}
\title{Deprecated: Get dataset documentation}
\usage{
get_dataset_doc(dataset_id, dest_dir = ".", download = TRUE)
}
\arguments{
\item{dataset_id}{dataset ID. See \code{sp_datasets}.}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}

\item{download}{Whether to download (the default) or open link in browser.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Deprecated, use \code{sp_get_dataset_doc()} instead.\cr\cr
\lifecycle{deprecated}
}
