% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_table}
\alias{get_table}
\title{Get a statnipokladna table}
\usage{
get_table(
  table_id,
  year = 2018,
  month = 12,
  ico = NULL,
  force_redownload = FALSE
)
}
\arguments{
\item{table_id}{A table ID. See \code{id} column in \code{sp_tables} for a list of available codelists.}

\item{year}{year, numeric, 2015-2018 for some datasets, 2010-2018 for others. Can be a vector of length > 1 (see details).}

\item{month}{month, numeric. Must be 3, 6, 9 or 12. Can be a vector of length > 1 (see details).}

\item{ico}{ID(s) of org to return, character of length one or more. If unset, returns all orgs. ID not checked for correctness/existence. See \url{http://monitor.statnipokladna.cz/2019/zdrojova-data/prohlizec-ciselniku/ucjed} to look up ID of any org in the dataset.}

\item{force_redownload}{Redownload even if recent file present? Defaults to FALSE.}
}
\value{
a tibble
}
\description{
Cleans and loads a table. If needed, a dataset containing the table is downloaded.
}
\details{
The data is loaded from files downloaded automatically by \code{get_dataset()};
files persist in a temporary directory per session.

Data is processed in the following way:
\itemize{
\item all columns are given names that are human-readable and facilitato add codelists
\item ICO (org. IDs) are normalised as in some datasets they are padded with leading zeros
\item a period, per_yr and per_m columns are created to identify the time period
\item value columns (named "ZU_") are transformed into numerics
\item other columns are left as character to avoid losing information
}
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Core workflow: 
\code{\link{add_codelist}()},
\code{\link{get_codelist}()},
\code{\link{get_dataset}()}
}
\concept{Core workflow}
