% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_table}
\alias{get_table}
\title{Deprecated: Get a statnipokladna table\cr\cr
Deprecated, use \code{sp_get_table()} instead.}
\usage{
get_table(
  table_id,
  year = 2018,
  month = 12,
  ico = NULL,
  redownload = FALSE,
  dest_dir = NULL
)
}
\arguments{
\item{table_id}{A table ID. See \code{id} column in \code{sp_tables} for a list of available codelists.}

\item{year}{year, numeric, 2015-2018 for some datasets, 2010-2018 for others.
Can be a vector of length > 1 (see Details for how to work with data across time periods.).}

\item{month}{month, numeric. Must be 3, 6, 9 or 12. Can be a vector of length > 1 (see details).}

\item{ico}{ID(s) of org to return, character of length one or more. If unset, returns all orgs. ID not checked for correctness/existence. See \url{http://monitor.statnipokladna.cz/datovy-katalog/prohlizec-ciselniku/ucjed} to look up ID of any org in the dataset.}

\item{redownload}{Redownload even if recent file present? Defaults to FALSE.}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\lifecycle{deprecated}
}
