% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\docType{data}
\name{sp_codelists}
\alias{sp_codelists}
\title{List of available codelists}
\format{A data frame with 27 rows and 2 variables:
\describe{
\item{\code{id}}{character. ID, used as \code{codelist_id} argument in \code{get_codelist}.}
\item{\code{name}}{character. Short name, mostly corresponds to title used on statnipokladna.cz.}
}}
\usage{
sp_codelists
}
\description{
Contains IDs and names of all available codelists that can be retrieved by get_codelist.
See \url{https://monitor.statnipokladna.cz/2019/zdrojova-data/ciselniky} for a more detailed
descriptions and a GUI for exploring the lists.
}
\details{
The \code{id} is to be used as the \code{codelist_id} parameter in \code{get_codelist}.
}
\keyword{datasets}
