% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sp_get_dataset_doc}
\alias{sp_get_dataset_doc}
\title{Get dataset documentation}
\usage{
sp_get_dataset_doc(dataset_id, dest_dir = NULL, download = TRUE)
}
\arguments{
\item{dataset_id}{dataset ID. See \code{sp_datasets}.}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{statnipokladna.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}

\item{download}{Whether to download (the default) or open link in browser.}
}
\value{
(invisible) path to file if \code{download = TRUE}, URL otherwise
}
\description{
Downloads XLS file with dataset documentation, or opens link to this file in browser.
}
\examples{
\donttest{
sp_get_dataset_doc("finm")
}
}
\seealso{
Other Utilities: 
\code{\link{get_dataset}()},
\code{\link{sp_get_codelist_viewer}()}
}
\concept{Utilities}
