% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{sp_tables}
\alias{sp_tables}
\title{List of available tables (PARTIAL)}
\format{
A data frame with 2 rows and 4 variables:
\describe{
\item{\code{id}}{character Table id, used as \code{table_id} argument to \code{sp_get_table}.}
\item{\code{dataset_id}}{integer Table number.}
\item{\code{czech_name}}{character Czech name of the table.}
\item{\code{note}}{character Note.}
}
}
\usage{
sp_tables
}
\description{
Contains IDs and names of all available tables that can be
retrieved by sp_get_table. Look inside the XLS documentation for each dataset at \url{https://monitor.statnipokladna.gov.cz/datovy-katalog/transakcni-data}
to see more detailed descriptions. Note that tables do not correspond to the tabulka/\code{vtab} attribute of the tables, they represent files inside datasets.
}
\seealso{
Other Lists of available entities: 
\code{\link{sp_codelists}},
\code{\link{sp_datasets}}
}
\concept{Lists of available entities}
\keyword{datasets}
