% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnipokladna-package.R
\docType{package}
\name{statnipokladna-package}
\alias{statnipokladna}
\alias{statnipokladna-package}
\title{statnipokladna: Use Data from the Czech Public Finance Database}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Get programmatic access to data from the Czech public
    budgeting and accounting database, Státní pokladna.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petrbouchal/statnipokladna}
  \item Report bugs at \url{https://github.com/petrbouchal/statnipokladna/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
